webnn: Use int16_t to compute TFLite padding
Internally TFLite will use int16_t for padding calculations on operators
like conv2d and pool2d. Use int16_t in our own calculations and reject
the graph if the developer has requested more padding than can be
represented with this type.

(cherry picked from commit 8859cdd1408cc480a5931cd42adbe25915e59b38)

Fixed: 494341337
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7690057
Commit-Queue: Phillis Tang <phillis@chromium.org>
Auto-Submit: Reilly Grant <reillyg@chromium.org>
Cr-Original-Commit-Position: refs/heads/main@{#1602922}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7699077
Commit-Queue: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>
Bot-Commit: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>
Commit-Queue: Reilly Grant <reillyg@chromium.org>
Cr-Commit-Position: refs/branch-heads/7727@{#1428}
Cr-Branched-From: ce01102937348db7b88c8a4257ee4b3ac702eb1a-refs/heads/main@{#1596535}

+++ services/webnn/tflite/graph_builder_tflite.h
@@ -490,13 +490,13 @@ class GraphBuilderTflite final {
                                          TensorIndex false_tensor_index,
                                          TensorIndex output_tensor_index);
 
-  // Insert a tempary pad operation if the `paddings` can't be converted to
+  // Insert a temporary pad operation if the `paddings` can't be converted to
   // tflite padding mode.
   base::expected<TensorIndex, std::string> InsertPadOperation(
       const TensorInfo& input_tensor_info,
-      base::span<const uint32_t> paddings);
+      base::span<const int16_t> paddings);
 
-  // Insert a tempary transpose operation for input operand with calling
+  // Insert a temporary transpose operation for input operand with calling
   // `SerializeTransposeOperation`.
   base::expected<TensorIndex, std::string> InsertTransposeOperation(
       const TensorInfo& input_tensor_info,
